+++ kubernetes-client/src/main/java/io/fabric8/kubernetes/client/dsl/FunctionCallable.java
@@ -16,7 +16,7 @@
 package io.fabric8.kubernetes.client.dsl;
 
 
-import io.fabric8.kubernetes.api.builder.Function;
+import java.util.function.Function;
 
 public interface FunctionCallable<I> {