+++ kubernetes-client/src/main/java/io/fabric8/kubernetes/client/dsl/internal/apps/v1/StatefulSetOperationsImpl.java
@@ -15,6 +15,7 @@
  */
 package io.fabric8.kubernetes.client.dsl.internal.apps.v1;
 
+import io.fabric8.kubernetes.api.builder.Visitor;
 import io.fabric8.kubernetes.api.model.Container;
 import io.fabric8.kubernetes.api.model.HasMetadata;
 import io.fabric8.kubernetes.api.model.Pod;
@@ -22,6 +23,7 @@ import io.fabric8.kubernetes.api.model.Status;
 import io.fabric8.kubernetes.api.model.apps.ControllerRevision;
 import io.fabric8.kubernetes.api.model.apps.ControllerRevisionList;
 import io.fabric8.kubernetes.api.model.apps.StatefulSet;
+import io.fabric8.kubernetes.api.model.apps.StatefulSetBuilder;
 import io.fabric8.kubernetes.api.model.apps.StatefulSetList;
 import io.fabric8.kubernetes.api.model.extensions.DeploymentRollback;
 import io.fabric8.kubernetes.client.Config;
@@ -257,6 +259,11 @@ public class StatefulSetOperationsImpl extends RollableScalableResourceOperation
     ControllerRevision previousControllerRevision = controllerRevisions.get(1);
 
     return sendPatchedStatefulSetData(previousControllerRevision.getData());
+  }
+
+  @Override
+  public StatefulSet edit(Visitor... visitors) {
+    return patch(new StatefulSetBuilder(getMandatory()).accept(visitors).build());
   }
 
   private StatefulSet sendPatchedStatefulSet(Map<String, Object> patchedUpdate) {