+++ kubernetes-client/src/main/java/io/fabric8/kubernetes/client/dsl/internal/extensions/v1beta1/ReplicaSetOperationsImpl.java
@@ -15,11 +15,13 @@
  */
 package io.fabric8.kubernetes.client.dsl.internal.extensions.v1beta1;
 
+import io.fabric8.kubernetes.api.builder.Visitor;
 import io.fabric8.kubernetes.api.model.Container;
 import io.fabric8.kubernetes.api.model.Pod;
 import io.fabric8.kubernetes.api.model.Status;
 import io.fabric8.kubernetes.api.model.extensions.DeploymentRollback;
 import io.fabric8.kubernetes.api.model.extensions.ReplicaSet;
+import io.fabric8.kubernetes.api.model.extensions.ReplicaSetBuilder;
 import io.fabric8.kubernetes.api.model.extensions.ReplicaSetList;
 import io.fabric8.kubernetes.client.Config;
 import io.fabric8.kubernetes.client.KubernetesClientException;
@@ -232,6 +234,11 @@ public class ReplicaSetOperationsImpl extends RollableScalableResourceOperation<
   @Override
   public Loggable<LogWatch> withLogWaitTimeout(Integer logWaitTimeout) {
     return new ReplicaSetOperationsImpl(((RollingOperationContext) context), logWaitTimeout);
+  }
+
+  @Override
+  public ReplicaSet edit(Visitor... visitors) {
+    return patch(new ReplicaSetBuilder(getMandatory()).accept(visitors).build());
   }
 
   static Map<String, String> getReplicaSetSelectorLabels(ReplicaSet replicaSet) {