+++ openshift-client/src/main/java/io/fabric8/openshift/client/dsl/internal/GroupOperationsImpl.java
@@ -15,10 +15,12 @@
  */
 package io.fabric8.openshift.client.dsl.internal;
 
+import io.fabric8.kubernetes.api.builder.Visitor;
 import io.fabric8.kubernetes.client.dsl.Resource;
 import io.fabric8.kubernetes.client.dsl.base.OperationContext;
 import okhttp3.OkHttpClient;
 import io.fabric8.openshift.api.model.Group;
+import io.fabric8.openshift.api.model.GroupBuilder;
 import io.fabric8.openshift.api.model.GroupList;
 import io.fabric8.openshift.client.OpenShiftConfig;
 
@@ -43,6 +45,11 @@ public class GroupOperationsImpl extends OpenShiftOperation<Group, GroupList, Re
   @Override
   public GroupOperationsImpl newInstance(OperationContext context) {
     return new GroupOperationsImpl(context);
+  }
+
+  @Override
+  public Group edit(Visitor... visitors) {
+    return patch(new GroupBuilder(getMandatory()).accept(visitors).build());
   }
 
   @Override